change file names and change rendering for uploader section
