+++ frontend/pages/ManageControlsPage/SetupExperience/cards/SetupAssistant/_styles.scss
@@ -7,6 +7,16 @@
     gap: $pad-xxlarge;
   }
 
+  &__upload-container {
+    display: flex;
+    flex-direction: column;
+    gap: $pad-large;
+  }
+
+  &__section-description {
+    margin: 0;
+  }
+
   @media (max-width: $break-md) {
     &__content {
       flex-direction: column;
