--- frontend/pages/ManageControlsPage/SetupExperience/cards/SetupAssistant/components/SetupAssistantPackageUploader/_styles.scss
@@ -1,6 +1,0 @@
-.setup-assistant-package-uploader {
-  > p {
-    font-size: $x-small;
-    margin: 0 0 $pad-large;
-  }
-}
