--- frontend/pages/ManageControlsPage/SetupExperience/cards/SetupAssistant/components/SetupAssistantPackageUploader/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./SetupAssistantPackageUploader";
