+++ frontend/pages/ManageControlsPage/SetupExperience/cards/SetupAssistant/components/SetupAssistantProfileCard/SetupAssistantProfileCard.tsx
@@ -1,0 +1,98 @@
+import React from "react";
+
+import URL_PREFIX from "router/url_prefix";
+import { uploadedFromNow } from "utilities/date_format";
+import endpoints from "utilities/endpoints";
+
+import Icon from "components/Icon";
+import Card from "components/Card";
+import Graphic from "components/Graphic";
+import Button from "components/buttons/Button";
+
+const baseClass = "setup-assistant-profile-card";
+
+interface ITestFormProps {
+  url: string;
+  token: string;
+  className?: string;
+}
+
+/**
+ * This component abstracts away the downloading of the package. It implements this
+ * with a browser form that calls the correct url to initiate the package download.
+ * We do it this way as this allows us to take advantage of the browsers native
+ * downloading UI instead of having to handle this in the Fleet UI.
+ * TODO: make common component and use here and in DownloadInstallers.tsx.
+ */
+const DownloadPackageButton = ({ url, token, className }: ITestFormProps) => {
+  return (
+    <form
+      key="form"
+      method="GET"
+      action={url}
+      target="_self"
+      className={className}
+    >
+      <input type="hidden" name="token" value={token || ""} />
+      <Button
+        variant="text-icon"
+        type="submit"
+        className={`${baseClass}__download-button`}
+      >
+        <Icon name="download" />
+      </Button>
+    </form>
+  );
+};
+
+interface ISetupAssistantProfileCardProps {
+  profileMetaData: any;
+  currentTeamId: number;
+  onDelete: (packageMetaData: any) => void;
+}
+
+const SetupAssistantProfileCard = ({
+  profileMetaData,
+  currentTeamId,
+  onDelete,
+}: ISetupAssistantProfileCardProps) => {
+  profileMetaData = {
+    title: "test-profile.json",
+    created_at: "2021-08-25T20:00:00Z",
+    token: "123-abc",
+  };
+
+  const { origin } = global.window.location;
+  const path = `${endpoints.MDM_BOOTSTRAP_PACKAGE}`;
+  const url = `${origin}${URL_PREFIX}/api${path}`;
+
+  return (
+    <Card paddingSize="medium" className={baseClass}>
+      <Graphic name="file-configuration-profile" />
+      <div className={`${baseClass}__info`}>
+        <span className={`${baseClass}__profile-name`}>
+          {profileMetaData.title}
+        </span>
+        <span className={`${baseClass}__uploaded-at`}>
+          {uploadedFromNow(profileMetaData.created_at)}
+        </span>
+      </div>
+      <div className={`${baseClass}__actions`}>
+        <DownloadPackageButton
+          className={`${baseClass}__download-package`}
+          url={url}
+          token={profileMetaData.token}
+        />
+        <Button
+          className={`${baseClass}__delete-button`}
+          variant="text-icon"
+          onClick={() => onDelete(profileMetaData)}
+        >
+          <Icon name="trash" color="ui-fleet-black-75" />
+        </Button>
+      </div>
+    </Card>
+  );
+};
+
+export default SetupAssistantProfileCard;
