+++ frontend/pages/ManageControlsPage/SetupExperience/cards/SetupAssistant/components/SetupAssistantProfileCard/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./SetupAssistantProfileCard";
