--- frontend/pages/ManageControlsPage/SetupExperience/cards/SetupAssistant/components/SetupAssistantUploadedProfileView/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./SetupAssistantUploadedProfileView";
